+增加了独立的Region对象创建接口 *修改了GrapX CLI引用clstd的配置

+++ GrapX/User/Win32Emu/GXUser_ExDC.cpp
@@ -337,7 +337,7 @@ GXHBITMAP GXDLLAPI gxLoadBitmapW(
     {
       GXLPCWSTR szBaseDir = _CLTEXT("Resource\\OBM\\");
       GXLPCWSTR szOBMFile = NULL;
-      switch((GXDWORD)lpBitmapName)
+      switch((GXDWORD)(size_t)lpBitmapName)
       {
       case GXOBM_CLOSE:
         break;
